BackEnd/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "reportlab>=4.4.4",
]
